Image module as a seperate file
